fix(i18n): load shared helpers before bundles; keep locale list bundle-independent — i18n_shared.js (parser-blocking) provides the _i18n* helpers that bundle VALUE references need at evaluation time; KNOWN_LOCALES decouples resolution and the language dropdown from loaded bundles so preferences are never clobbered to 'en'
